requirements/prod.txt
# ====================
# Web Framework
# ====================
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ====================
# Database
# ====================
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0          # async PostgreSQL driver
psycopg2-binary==2.9.9   # sync PostgreSQL driver (used by Alembic)

# ====================
# Authentication & Security
# ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ====================
# Settings & Environment
# ====================
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ====================
# Redis & Task Queue
# ====================
redis==5.0.3
celery==5.3.6
flower==2.0.1             # Celery monitoring (optional but useful in prod)

# ====================
# HTTP Client
# ====================
httpx==0.27.0

# ====================
# File Handling
# ====================
python-magic==0.4.27
aiofiles==23.2.1
Pillow==10.3.0
# zipstream==1.5.2  <-- removed as it's invalid and redundant with zipstream-new below

# ====================
# Email
# ====================
fastapi-mail==1.4.1

# ====================
# Logging & Monitoring
# ====================
loguru==0.7.2

# ====================
# Utilities
# ====================
pytz==2024.1
python-dateutil==2.9.0
zipstream-new==1.1.8
